personal: add some consts(啦啦啦,不写代码啦,出去吃饭啦,单身狗的日常~)

+++ lib-blog-db/src/main/java/coder/blog/db/article/BlogArticle.java
@@ -34,7 +34,7 @@ public class BlogArticle extends BaseEntity{
 
     // 原创、转载、翻译
     @Column(name = "type")
-    private Integer type;
+    private Short type;
 
     // 原始地址: 未保证版权, 转载、翻译的文章需注明原文链接
     @Column(name = "original_address")
